requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
faiss-cpu
pypdf